requirements.txt
fastapi
uvicorn
google-cloud-storage
pandas
pytz
matplotlib